move html and css validator into viz utils to prevent circular dependencies that are causing issues in python 27
